+++ lib/ogr/geometry/interfaces.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-module OGR
-  class Geometry
-    # A module that contains various interfaces, used by different geometry types.
-    module Interfaces
-    end
-  end
-end