+++ lib/ogr/geometry_collection.rb
@@ -1,11 +1,13 @@
 # frozen_string_literal: true
 
-require_relative 'geometry_types/container'
+require_relative 'geometry/container'
+require_relative 'geometry/polygon_from_edges'
 
 module OGR
   class GeometryCollection < OGR::Geometry
-    include OGR::GeometryTypes::Container
     include GDAL::Logger
+    include OGR::Geometry::Container
+    include OGR::Geometry::PolygonFromEdges
 
     GEOMETRY_TYPE = :wkbGeometryCollection