+++ lib/ogr/polygon.rb
@@ -1,13 +1,13 @@
 # frozen_string_literal: true
 
-require_relative 'geometry_types/container'
-require_relative 'geometry/interfaces/area'
+require_relative 'geometry/container'
+require_relative 'geometry/area'
 
 module OGR
   class Polygon < OGR::Geometry
-    include OGR::GeometryTypes::Container
-    include OGR::Geometry::Interfaces::Area
     include GDAL::Logger
+    include OGR::Geometry::Container
+    include OGR::Geometry::Area
 
     GEOMETRY_TYPE = :wkbPolygon