+++ Changelog.md
@@ -84,13 +84,14 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
   `CoordinateTranformation`. In < 3, the Ruby code checks the returned pointer and raises if it's
   null. Now these raise the same exception.
 - _BREAKING_: `OGR::Driver.create_data_source` raises on failure instead of returning `nil`.
-- _BREAKING_: Moved `OGR::Geometry#point_count` to `OGR::Geometry::Interfaces::XYPoints` and
-  `XYZPoints`.
+- _BREAKING_: Moved `OGR::Geometry#point_count` to `OGR::Geometry::XYPoints` and `XYZPoints`.
 - _BREAKING_: Moved `OGR::Geometry#geometry_count` to `OGR::Geometry::Container`.
+- _BREAKING_: Moved `OGR::Geometry::Container#polygon_from_edges` to
+  `OGR::Geometry::PolygonFromEdges`.
 - `OGR::SpatialReferenceMixins::Importers` now return `self` instead of the `OGR` error code
   `Symbol`.
 - Extracted methods from `OGR::GeometryTypes::Curve` and `OGR::GeometryTypes::Surface` to new
-  `OGR::Geometry::Interfaces` modules: `XYPoints`, `XYZPoints`, `Area`, `Length`.
+  `OGR::Geometry` mixins: `XYPoints`, `XYZPoints`, `Area`, `Length`.
 - `OGR::Geometry` is now a (abstract) `class` from which other, concrete geometry objects inherit.
 
 ### Fixed